+++ src/fileManagement/fileManagement.js
@@ -1,98 +1,5 @@
-const path = require("path");
 const fs = require("fs");
-const glob = require('glob');
-const vscode = require('vscode');
-const constants = require("../constants");
-
-//#region AL File Creation
-/**
- * 
- * @param {string} destinationPath 
- * @param {string} objectType 
- * @param {number} objectId 
- * @param {string} objectName 
- * @param {string} objectNamePrefix 
- * @param {(objectType: string, objectId: number, objectName: string) => string} fileNameFormatter 
- */
-function createAlFile(destinationPath, objectType, objectId, objectName, objectNamePrefix, fileNameFormatter){
-    const newObjectId = objectId + 80000;
-    let newObjectName = objectName.replace(/[^\w]/g,'');
-    let fileContent;
-    // Generate File Name and Content
-    switch (objectType) {
-        case constants.AlObjectTypes.tableExtension:
-            fileContent = 
-        `
-{
-    fields
-    {
-
-    }
-}`;
-            break;
-        case constants.AlObjectTypes.pageExtension:
-            fileContent = 
-        `
-{
-    layout
-    {
-    }
-
-    actions
-    {
-    }
-}`;
-            break;
-        default:
-            console.error('Unsuported AL object type.');
-            throw Error('Unsuported AL object type.');
-    }
-    
-    const fileName = fileNameFormatter(objectType, objectId, objectName);
-    //vscode.window.showInformationMessage(fileName);
-    //Generate Path
-    const filePath = path.join(destinationPath, fileName);
-    return existsFileAsync(filePath)
-            .then(exists => {
-            if (exists) {
-                return filePath;
-            }
-            const contents = objectType + ' ' + newObjectId + ' "' + objectNamePrefix + newObjectName + '" extends "' +  objectName + '" //' + objectId + fileContent;
-            return writeTextFileAsync(filePath, contents)
-                .then(() => filePath);
-            });  
-}
-exports.createAlFile = createAlFile;
-
-/**
- * 
- * @param {string} objectType 
- * @param {number} objectId 
- * @param {string} objectName
- * @returns {string}
- */
-function oldAlFileNameFormatter(objectType, objectId, objectName){
-    const newObjectId = 80000 + objectId;
-    const prefix = constants.AlObjectTypesToFilePrefix(objectType);
-    const newObjectName = objectName.replace(/[^\w]/g,'');
-    return prefix + objectId + '-' + 'Ext' + newObjectId + '.' + newObjectName + '.al';
-}
-exports.oldAlFileNameFormatter = oldAlFileNameFormatter;
-
-/**
- * 
- * @param {string} objectType 
- * @param {number} objectId 
- * @param {string} objectName
- * @returns {string}
- */
-function newAlFileNameFormatter(objectType, objectId, objectName){
-    const newObjectName = objectName.replace(/[^\w]/g,'');
-    const fullTypeName = constants.AlObjectTypesToFullTypeName(objectType);
-    return newObjectName + '.' + fullTypeName + '.al';
-}
-exports.newAlFileNameFormatter = newAlFileNameFormatter;
-//#endregion
+const path = require("path");
 
 //#region File Creation
 function writeTextFileAsync(filename, data) {
@@ -105,6 +12,7 @@ function writeTextFileAsync(filename, data) {
         });
     });
 }
+exports.writeTextFileAsync = writeTextFileAsync;
 function existsFileAsync(filename) {
     return new Promise(resolve => {
         fs.exists(filename, exists => {
@@ -112,42 +20,7 @@ function existsFileAsync(filename) {
         });
     });
 }
-//#endregion
-
-//#region Workspace Finder
-function getCurrentWorkspaceFolderPath() {
-    const currentWorkspaceFolder = getCurrentWorkspaceFolder();
-    if (currentWorkspaceFolder) {
-        return currentWorkspaceFolder.uri.fsPath;
-    }
-    return undefined;
-}
-exports.getCurrentWorkspaceFolderPath = getCurrentWorkspaceFolderPath;
-
-function getCurrentWorkspaceFolder() {
-    const activeEditor = vscode.window.activeTextEditor;
-    if (activeEditor) {
-        if (activeEditor.document) {
-            lastActiveWorkspace = getWorkspaceFolderFromPath(activeEditor.document.fileName);
-            return lastActiveWorkspace;
-        }
-    }
-}
-
-function getWorkspaceFolderFromPath(path) {
-    if (!Array.isArray(vscode.workspace.workspaceFolders) || vscode.workspace.workspaceFolders.length === 0) {
-        lastActiveWorkspace = undefined;
-        return lastActiveWorkspace;
-    }
-    // one workspace folder, this is the old folder/rootpath scenario.
-    if (vscode.workspace.workspaceFolders.length === 1) {
-        lastActiveWorkspace = vscode.workspace.workspaceFolders[0];
-        return lastActiveWorkspace;
-    }
-    const uri = vscode.Uri.file(path);
-    const workspace = vscode.workspace.getWorkspaceFolder(uri);
-    return workspace;
-}
+exports.existsFileAsync = existsFileAsync;
 //#endregion
 
 //#region Folder Creation
@@ -160,47 +33,3 @@ function createFolder(dir) {
 }
 exports.createFolder = createFolder;
 //#endregion
-
-//#region Open AL files
-/**
- * @param {string} objectName
- * @param {string} alObjectType
- * @returns {Thenable<vscode.TextDocument>[]} 
- */
-function openALFile(objectName, alObjectType) {
-    const tenables = [];
-    const files = getAlFileLocations(objectName, alObjectType);
-    files.forEach(file => tenables.push(
-        vscode.workspace.openTextDocument(file).then(doc => {
-            vscode.window.showTextDocument(doc, {preserveFocus: true, preview: false});
-        })
-    ));
-    return tenables;
-}
-exports.openALFile = openALFile;
-
-/**
- * @param {string} objectName
- * @param {string} alObjectType
- * @returns {string[]} File locations 
- */
-function getAlFileLocations(objectName, alObjectType){
-    const compactObjectName = objectName.replace(/[^\w]/g, '');
-
-    const UseOldFileNamingConventions = vscode.workspace.getConfiguration('ALTB').get('UseOldFileNamingConventions');
-    
-    let fileLocationFormat = `${getCurrentWorkspaceFolderPath()}/src/**/`;
-    if(UseOldFileNamingConventions) {
-        const filePrefix = constants.AlObjectTypesToFilePrefix(alObjectType);
-        fileLocationFormat += `${filePrefix}*${compactObjectName}.al`;
-    } else {
-        const fullTypeName = constants.AlObjectTypesToFullTypeName(alObjectType);
-        fileLocationFormat += `${compactObjectName}.${fullTypeName}.al`;
-    }
-
-    return glob.sync(fileLocationFormat);
-}
-exports.getAlFileLocations = getAlFileLocations;
-//#endregion
-
-let lastActiveWorkspace = undefined;
