+++ src/relatedTables/relatedTables.js
@@ -1,6 +1,8 @@
 const vscode = require('vscode');
 const constants = require('../constants');
 const fileMangagement = require('../fileManagement/fileManagement');
+const alFileMangagement = require('../fileManagement/alFileManagement');
+const workspaceManagement = require('../fileManagement/workspaceManagement');
 
 /**
  * @param {string} objectNamePrefix
@@ -8,7 +10,7 @@ const fileMangagement = require('../fileManagement/fileManagement');
  * @returns {Promise}
  */
 exports.createRelatedTables = function createRelatedTables(objectNamePrefix, fileNameFormatter) {
-    const rootPath = fileMangagement.getCurrentWorkspaceFolderPath();
+    const rootPath = workspaceManagement.getCurrentWorkspaceFolderPath();
     // if rootpath is empty then error
     if (!rootPath) {
         return Promise.reject("No AL workspace folder is active");
@@ -26,7 +28,7 @@ exports.createRelatedTables = function createRelatedTables(objectNamePrefix, fil
         fileMangagement.createFolder(destinationPath);
 
         element.objects.forEach(object => {
-            fileCreationPromises.push(fileMangagement.createAlFile(destinationPath, element.objectType, object.id, object.name, objectNamePrefix, fileNameFormatter));
+            fileCreationPromises.push(alFileMangagement.createAlFile(destinationPath, element.objectType, object.id, object.name, objectNamePrefix, fileNameFormatter));
         });
     });
 
@@ -112,7 +114,7 @@ exports.RelatedTablesManager = class RelatedTablesManager {
     openRelateObjects(objectName, sourceAlObjectType, alObjectTypes) {
         const relatedObjects = this.getRelateObjects(objectName, sourceAlObjectType, alObjectTypes);
         relatedObjects.forEach(relatedObject =>
-            fileMangagement.openALFile(relatedObject.name, relatedObject.type)
+            alFileMangagement.openALFile(relatedObject.name, relatedObject.type)
         );
         return relatedObjects.length > 0;
     }
